backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
newspaper3k
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
transformers